planning readme part 1
